backend/requirements.txt
pytest==6.2.5
pytest-cov==4.0.0
python-dotenv==0.21.0
functions-framework==3.*
pandas
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.12
langchain-text-splitters
langchain-huggingface
transformers 
sentence_transformers 
faiss-cpu